requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-environ==0.8.1
Django-Verify-Email==1.0.9
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
openpyxl==3.0.9
psycopg2-binary==2.9.3
pymemcache==3.5.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
